refactor(net): split runtime state roles
